+++ NEWS.md
@@ -1,5 +1,6 @@
 # ggplot2 (development version)
 
+* New stat: `stat_manual()` for arbitrary computations (@teunbrand, #3501)
 * Reversal of a dimension, typically 'x' or 'y', is now controlled by the 
   `reverse` argument in `coord_cartesian()`, `coord_fixed()`, `coord_radial()`
   and `coord_sf()`. In `coord_radial()`, this replaces the older `direction` 
@@ -237,6 +238,7 @@
   (@teunbrand, #4722, #6069).
 * `geom_abline()` clips to the panel range in the vertical direction too
   (@teunbrand, #6086).
+* Added `panel.widths` and `panel.heights` to `theme()` (#5338, @teunbrand).
 
 # ggplot2 3.5.1